[FP-76] Refactor: api 함수 파일 정리 및 리뷰 리스트 컴포넌트 리팩토링
