+++ src/apis/ActivityDetailsPage/getActivityAvailableSchedule.ts
@@ -1,6 +1,17 @@
-import { getActivityAvailableSchedule } from '@/apis/ActivityDetailsPage/activityDetails';
+import { axiosRequester } from '@/libs/axios';
 import { AvailableScheduleTypeProps, DaySchedule } from '@/types/availableSchedulesTypes';
 import { useQuery } from '@tanstack/react-query';
+
+export const getActivityAvailableSchedule = async ({ activityId, year, month }: AvailableScheduleTypeProps) => {
+  const { data } = await axiosRequester<DaySchedule[]>({
+    options: {
+      method: 'GET',
+      url: `/activities/${activityId}/available-schedule?year=${year}&month=${month}`,
+    },
+  });
+
+  return data;
+};
 
 export const useActivityAvailableSchedule = ({ activityId, year, month }: AvailableScheduleTypeProps) => {
   const {
