+++ src/apis/ActivityDetailsPage/postActivityReservation.ts
@@ -1,6 +1,19 @@
-import { postActivityReservation } from '@/apis/ActivityDetailsPage/activityDetails';
+import { axiosRequester } from '@/libs/axios';
 import { formSubmitDataAtomType } from '@/store/activityDetailsAtom';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+export const postActivityReservation = async ({ activityId, scheduleId, headCount }: formSubmitDataAtomType) => {
+  const { data } = await axiosRequester({
+    options: {
+      method: 'POST',
+      url: `/activities/${activityId}/reservations`,
+      data: { scheduleId, headCount },
+    },
+    includeAuth: true,
+  });
+
+  return data;
+};
 
 export const useActivityReservationMutation = () => {
   const queryClient = useQueryClient();
